r5312/sofa-dev : FIX: change unsigned to size_t in mycuda.h:mycudaMemset() mycudaMemset declaration and definition were different. (.h and .cu)

+++ modules/sofa/gpu/cuda/mycuda.h
@@ -58,7 +58,7 @@ extern "C" {
     extern void SOFA_GPU_CUDA_API mycudaGLMapBufferObject(void** ptr, int id);
     extern void SOFA_GPU_CUDA_API mycudaGLUnmapBufferObject(int id);
 
-    extern void SOFA_GPU_CUDA_API mycudaMemset(void * devPtr, int val , unsigned size);
+    extern void SOFA_GPU_CUDA_API mycudaMemset(void * devPtr, int val , size_t size);
 
     extern void SOFA_GPU_CUDA_API mycudaLogError(const char* err, const char* src);
     extern int myprintf(const char* fmt, ...);
